[kb] Remove create_with_new_index()
I changed `create_resin_index()` to be a method, which happens after KB instantiation.
We were checking for `if self._index in None` in all of our functions anyway - so this doesn't change the error logic.

If the user will create a KB instance with a name of a non-existing index, it will be printed that he has to create it first