[kb] Removed `index_params` from __init__()
It doesn't make too much sense to require the `index_params` argument in the constructor. In interactive usage - this param should be passed directly to the `create_resin_index()` method.
I had to complicate the code of `from_config()` a bit - but it makes the usage pattern of KnowledgeBase much more explicit and readable.