CLI and server bug fixes on Windows (#170)
* [cli] Bug fix - used `os.path.join()` instead of 'urljoin()'

That created an error on Windows

* [server] Support `canopy stop` on Windows

Two bugs:
1. Trying to `pgrep gunicorn` on Windows, even though windows doesn't have `pgrep` and Gunicorn isn't supported on windows.
2. In windows we need to send the `CTRL_C` signal, which has a totally different implmentation behind it than GITINT

* [cli] Improve chat warnning message

Since there is no Gunicorn on windows, it doesn't make sense to tell the users to run it

* [cli] Bug fix - CLI couldn't open server routes on Windows

The CLI used the url `http://0.0.0.0:8000` by default, which is not supported on windows.
I replaced it with `http://localhost:8000`.

* Don't use tenerary if, so mypy will be happy

What can you do...

* [cli] Bug fix - still using os.path.join() for url

Caught in PR #170 review.

* linter